+++ imas_muscle3/visualization/examples/simple_1d_plot/simple_1d_plot.ymmsl
@@ -5,19 +5,18 @@ model:
     source_component:
       implementation: source_component
       ports:
-        o_i: [equilibrium_out, pf_active_out]
+        o_i: [equilibrium_out]
     visualization_component:
       implementation: visualization_component
       ports:
-        s: [equilibrium_in, pf_active_in]
+        s: [equilibrium_in]
   conduits:
     source_component.equilibrium_out: visualization_component.equilibrium_in
-    source_component.pf_active_out: visualization_component.pf_active_in
 settings:
   source_component.source_uri: imas:hdf5?path=/home/sebbe/projects/iter_python/IMAS-MUSCLE3/data/striped_eq
-  visualization_component.plot_file_path: /home/sebbe/projects/iter_python/IMAS-MUSCLE3/imas_muscle3/visualization/examples/eq_and_pfa/eq_and_pfa.py
+  visualization_component.plot_file_path: /home/sebbe/projects/iter_python/IMAS-MUSCLE3/imas_muscle3/visualization/examples/simple_1d_plot/simple_1d_plot.py
   visualization_component.port: 5007
-  visualization_component.throttle_interval: 0.5
+  visualization_component.throttle_interval: 0
   visualization_component.keep_alive: true
 implementations:
   visualization_component:
